requirements.txt
# Core Deep Learning Libraries
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer Vision and Image Processing
opencv-python>=4.6.0
Pillow>=9.2.0
scikit-image>=0.19.0

# Data Augmentation
albumentations>=1.3.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning Utilities
scikit-learn>=1.1.0

# Visualization and Plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress Bars and Utilities
tqdm>=4.64.0

# Logging and Configuration
tensorboard>=2.10.0
wandb>=0.13.0  # Optional: for advanced experiment tracking

# JSON and Data Handling
jsonschema>=4.0.0
PyYAML>=6.0  # For YAML configuration files

# Dataset Management
kagglehub>=0.2.0  # For automatic dataset downloads from Kaggle

# Image Quality Metrics
lpips>=0.1.4  # Learned Perceptual Image Patch Similarity

# Optional: Advanced optimizations
# onnx>=1.12.0  # For model export
# onnxruntime>=1.12.0  # For optimized inference
# tensorrt>=8.4.0  # NVIDIA optimization (requires CUDA)

# Development and Testing (Optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Jupyter for experimentation (Optional)
jupyter>=1.0.0
ipywidgets>=8.0.0

# For Composition
transformers
timm